+++ android/src/main/kotlin/dev/fluttercommunity/workmanager/WorkmanagerPlugin.kt
@@ -4,7 +4,6 @@ import android.content.Context
 import io.flutter.embedding.engine.plugins.FlutterPlugin
 import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.MethodChannel
-import io.flutter.plugin.common.PluginRegistry
 
 class WorkmanagerPlugin : FlutterPlugin {
 
@@ -29,25 +28,5 @@ class WorkmanagerPlugin : FlutterPlugin {
         methodChannel?.setMethodCallHandler(null)
         methodChannel = null
         workmanagerCallHandler = null
-    }
-
-    companion object {
-        var pluginRegistryCallback: PluginRegistry.PluginRegistrantCallback? = null
-
-        @JvmStatic
-        fun registerWith(registrar: PluginRegistry.Registrar) {
-            val plugin = WorkmanagerPlugin()
-            plugin.onAttachedToEngine(registrar.context(), registrar.messenger())
-            registrar.addViewDestroyListener {
-                plugin.onDetachedFromEngine()
-                false
-            }
-        }
-
-        @Deprecated(message = "Use the Android v2 embedding method.")
-        @JvmStatic
-        fun setPluginRegistrantCallback(pluginRegistryCallback: PluginRegistry.PluginRegistrantCallback) {
-            Companion.pluginRegistryCallback = pluginRegistryCallback
-        }
     }
 }